+++ Packages/com.github.kurotu.vrc-quest-tools/Editor/Menus/AvatarBuilderMenu.cs
@@ -1,0 +1,20 @@
+#if !VQT_HAS_NDMF
+using KRT.VRCQuestTools.Menus;
+using KRT.VRCQuestTools.Models;
+using UnityEditor;
+
+namespace KRT.VRCQuestTools.Menu
+{
+    /// <summary>
+    /// Menu for avatar builder.
+    /// </summary>
+    internal static class AvatarBuilderMenu
+    {
+        [MenuItem(VRCQuestToolsMenus.MenuPaths.ShowAvatarBuilder, false, (int)VRCQuestToolsMenus.MenuPriorities.ShowAvatarBuilder)]
+        private static void InitFromMenu()
+        {
+            EditorUtility.DisplayDialog(VRCQuestTools.Name, VRCQuestToolsSettings.I18nResource.FeatureRequiresNdmf, "OK");
+        }
+    }
+}
+#endif