+++ CHANGELOG.md
@@ -9,9 +9,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - [NDMF] `VQT Mesh Flipper` component to flip the mesh of the avatar.
+- [NDMF] `[NDMF] Manual Bake with Android Settings` menu to the right-click menu of the avatar.
+- [NDMF] Feature to select an avatar to upload in the `VQT Avatar Builder`.
 
 ### Changed
 - [NDMF] Check texture format after TexTransTool.
+- [NDMF] Change the inspector of `VQT Avatar Converter Settings` to prioritize non-destructive conversion.
+- [NDMF] Convert avatars after lilycalInventory in the transforming phase.
 
 ### Fixed
 - [NDMF] ArgumentNullException on play mode when the avatar has a newly created Pipeline Manager.